[1.5.0-beta7]支持添加酿造台配方,相关GUI待完善

+++ src/main/java/com/github/yufiriamazenta/craftorithm/recipe/RecipeManager.java
@@ -2,24 +2,26 @@ package com.github.yufiriamazenta.craftorithm.recipe;
 
 import com.github.yufiriamazenta.craftorithm.Craftorithm;
 import com.github.yufiriamazenta.craftorithm.CraftorithmAPI;
-import com.github.yufiriamazenta.craftorithm.recipe.custom.AnvilRecipe;
 import com.github.yufiriamazenta.craftorithm.recipe.custom.CustomRecipe;
+import com.github.yufiriamazenta.craftorithm.recipe.custom.PotionMixRecipe;
 import com.github.yufiriamazenta.craftorithm.util.ContainerUtil;
 import com.github.yufiriamazenta.craftorithm.util.LangUtil;
 import crypticlib.CrypticLib;
 import crypticlib.config.impl.YamlConfigWrapper;
+import crypticlib.platform.IPlatform;
 import crypticlib.util.FileUtil;
+import io.papermc.paper.potion.PotionMix;
 import org.bukkit.Bukkit;
 import org.bukkit.Keyed;
 import org.bukkit.NamespacedKey;
 import org.bukkit.configuration.file.YamlConfiguration;
 import org.bukkit.inventory.*;
+import org.bukkit.potion.PotionBrewer;
 
 import java.io.File;
 import java.util.*;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.CopyOnWriteArrayList;
-import java.util.concurrent.atomic.AtomicReference;
 
 public class RecipeManager {
 
@@ -29,7 +31,7 @@ public class RecipeManager {
     private static final Map<String, List<NamespacedKey>> recipeGroupMap;
     private static final Map<String, Integer> recipeSortIdMap;
     private static final Map<NamespacedKey, Boolean> recipeUnlockMap;
-    private static final Map<NamespacedKey, AnvilRecipe> anvilRecipeMap;
+    private static final Map<String, List<NamespacedKey>> potionMixGroupMap;
     private static final List<NamespacedKey> serverRecipeCache;
     private static final List<Recipe> removedRecipeRecycleBin;
     private static final String PLUGIN_RECIPE_NAMESPACE = "craftorithm";
@@ -42,8 +44,13 @@ public class RecipeManager {
         serverRecipeCache = new CopyOnWriteArrayList<>();
         recipeUnlockMap = new ConcurrentHashMap<>();
         recipeSortIdMap = new ConcurrentHashMap<>();
-        anvilRecipeMap = new ConcurrentHashMap<>();
         removedRecipeRecycleBin = new CopyOnWriteArrayList<>();
+
+        if (supportPotionMix()) {
+            potionMixGroupMap = new ConcurrentHashMap<>();
+        } else {
+            potionMixGroupMap = null;
+        }
     }
 
     public static void loadRecipeManager() {
@@ -85,11 +92,21 @@ public class RecipeManager {
                 YamlConfigWrapper configWrapper = recipeConfigWrapperMap.get(fileName);
                 YamlConfiguration config = configWrapper.config();
                 boolean multiple = config.getBoolean("multiple", false);
+                Recipe[] recipes;
                 if (multiple) {
-                    Recipe[] multipleRecipes = RecipeFactory.newMultipleRecipe(config, fileName);
-                    regRecipes(fileName, Arrays.asList(multipleRecipes), configWrapper);
+                    recipes = RecipeFactory.newMultipleRecipe(config, fileName);
                 } else {
-                    Recipe[] recipes = RecipeFactory.newRecipe(config, fileName);
+                    recipes = RecipeFactory.newRecipe(config, fileName);
+                }
+                if (!supportPotionMix()) {
+                    regRecipes(fileName, Arrays.asList(recipes), configWrapper);
+                    continue;
+                }
+
+                RecipeType recipeType = RecipeType.valueOf(config.getString("type", "shaped").toUpperCase());
+                if (recipeType.equals(RecipeType.POTION)) {
+                    regPotionMix(fileName, Arrays.asList(recipes), configWrapper);
+                } else {
                     regRecipes(fileName, Arrays.asList(recipes), configWrapper);
                 }
             } catch (Exception e) {
@@ -105,23 +122,30 @@ public class RecipeManager {
         for (Recipe recipe : recipes) {
             NamespacedKey recipeKey = getRecipeKey(recipe);
             recipeKeyList.add(getRecipeKey(recipe));
-            if (recipe instanceof CustomRecipe) {
-                //TODO
-                if (Objects.requireNonNull(((CustomRecipe) recipe).getRecipeType()) == RecipeType.ANVIL) {
-                    regAnvilRecipe((AnvilRecipe) recipe);
-                }
+            Bukkit.addRecipe(recipe);
+            boolean defUnlockCondition = Craftorithm.getInstance().getConfig().getBoolean("all_recipe_unlocked", false);
+            if (config.contains("unlock")) {
+                recipeUnlockMap.put(recipeKey, config.getBoolean("unlock", defUnlockCondition));
             } else {
-                Bukkit.addRecipe(recipe);
-                boolean defUnlockCondition = Craftorithm.getInstance().getConfig().getBoolean("all_recipe_unlocked", false);
-                if (config.contains("unlock")) {
-                    recipeUnlockMap.put(recipeKey, config.getBoolean("unlock", defUnlockCondition));
-                } else {
-                    recipeUnlockMap.put(recipeKey, defUnlockCondition);
-                }
+                recipeUnlockMap.put(recipeKey, defUnlockCondition);
             }
         }
         recipeGroupMap.put(recipeName, recipeKeyList);
         recipeSortIdMap.put(recipeName, config.getInt("sort_id", 0));
+    }
+
+    public static void regPotionMix(String potionMixName, List<Recipe> recipes, YamlConfigWrapper configWrapper) {
+        PotionBrewer potionBrewer = Bukkit.getPotionBrewer();
+        List<NamespacedKey> potionMixes = new ArrayList<>();
+        for (Recipe recipe : recipes) {
+            if (!(recipe instanceof PotionMixRecipe)) {
+                continue;
+            }
+            potionBrewer.addPotionMix(((PotionMixRecipe) recipe).potionMix());
+            potionMixes.add(((PotionMixRecipe) recipe).getKey());
+        }
+        potionMixGroupMap.put(potionMixName, potionMixes);
+        recipeSortIdMap.put(potionMixName, configWrapper.config().getInt("sort_id", 0));
     }
 
     private static void reloadOtherPluginsRecipes() {
@@ -171,12 +195,15 @@ public class RecipeManager {
             removeCraftorithmRecipe(key, false);
         });
 
+        if (supportPotionMix()) {
+            Bukkit.getPotionBrewer().resetPotionMixes();
+        }
+
         //先将已经删除的配方还原
         for (Recipe recipe : removedRecipeRecycleBin) {
             Bukkit.addRecipe(recipe);
         }
         removedRecipeRecycleBin.clear();
-        anvilRecipeMap.clear();
         recipeGroupMap.clear();
         recipeUnlockMap.clear();
         recipeSortIdMap.clear();
@@ -196,10 +223,20 @@ public class RecipeManager {
             return RecipeType.SMITHING;
         else if (recipe instanceof StonecuttingRecipe)
             return RecipeType.STONE_CUTTING;
-        else if (recipe instanceof AnvilRecipe)
-            return RecipeType.ANVIL;
+        else if (recipe instanceof PotionMixRecipe)
+            return RecipeType.POTION;
         else
             return RecipeType.UNKNOWN;
+    }
+
+    /**
+     * 删除酿造台配方
+     * @param potionMixKeys 酿造台配方的key
+     */
+    public static void removePotionMix(List<NamespacedKey> potionMixKeys) {
+        for (NamespacedKey potionMixKey : potionMixKeys) {
+            Bukkit.getPotionBrewer().removePotionMix(potionMixKey);
+        }
     }
 
     /**
@@ -238,16 +275,30 @@ public class RecipeManager {
     }
 
     public static boolean removeCraftorithmRecipe(String recipeName, boolean deleteFile) {
-        int i = removeRecipes(recipeGroupMap.getOrDefault(recipeName, new ArrayList<>()));
-        YamlConfigWrapper recipeConfig = recipeConfigWrapperMap.get(recipeName);
-        if (recipeConfig != null && deleteFile) {
-            recipeConfig.configFile().delete();
-        }
-        recipeConfigWrapperMap.remove(recipeName);
-        recipeGroupMap.remove(recipeName);
-        recipeSortIdMap.remove(recipeName);
+        if (!recipeGroupMap.containsKey(recipeName)) {
+            if (potionMixGroupMap.containsKey(recipeName)) {
+                removePotionMix(potionMixGroupMap.get(recipeName));
+                YamlConfigWrapper recipeConfig = recipeConfigWrapperMap.get(recipeName);
+                if (recipeConfig != null && deleteFile) {
+                    recipeConfig.configFile().delete();
+                }
+                recipeSortIdMap.remove(recipeName);
+                return true;
+            } else {
+                return false;
+            }
+        } else {
+            int i = removeRecipes(recipeGroupMap.getOrDefault(recipeName, new ArrayList<>()));
+            YamlConfigWrapper recipeConfig = recipeConfigWrapperMap.get(recipeName);
+            if (recipeConfig != null && deleteFile) {
+                recipeConfig.configFile().delete();
+            }
+            recipeConfigWrapperMap.remove(recipeName);
+            recipeGroupMap.remove(recipeName);
+            recipeSortIdMap.remove(recipeName);
 
-        return i > 0;
+            return i > 0;
+        }
     }
 
     public static boolean disableOtherPluginsRecipe(List<NamespacedKey> recipeKeys) {
@@ -324,27 +375,6 @@ public class RecipeManager {
         return recipeSortIdMap.getOrDefault(recipeName, 0);
     }
 
-    public static AnvilRecipe matchAnvilRecipe(ItemStack base, ItemStack addition) {
-        if (base == null || addition == null)
-            return null;
-        AtomicReference<AnvilRecipe> anvilRecipe = new AtomicReference<>();
-        anvilRecipeMap.forEach((key, recipe) -> {
-            if (recipe.checkSource(base, addition)) {
-                anvilRecipe.set(recipe);
-            }
-        });
-        return anvilRecipe.get();
-    }
-
-    public static void regAnvilRecipe(AnvilRecipe recipe) {
-        anvilRecipeMap.put(recipe.getKey(), recipe);
-    }
-
-//    private static void regPotionMixRecipe(PotionMix recipe) {
-//        //TODO
-//    }
-
-
 
     public static File getRecipeFileFolder() {
         return recipeFileFolder;
@@ -352,6 +382,10 @@ public class RecipeManager {
 
     public static Map<String, List<NamespacedKey>> getRecipeGroupMap() {
         return recipeGroupMap;
+    }
+
+    public static Map<String, List<NamespacedKey>> getPotionMixGroupMap() {
+        return potionMixGroupMap;
     }
 
     public static YamlConfigWrapper getRemovedRecipeConfig() {
@@ -370,9 +404,13 @@ public class RecipeManager {
         return recipeConfigWrapperMap;
     }
 
-    public static Map<NamespacedKey, AnvilRecipe> getAnvilRecipeMap() {return anvilRecipeMap;}
     public static Map<String, Integer> getRecipeSortIdMap() {
         return recipeSortIdMap;
+    }
+
+    public static boolean supportPotionMix() {
+        IPlatform.Platform platform = CrypticLib.platform().platform();
+        return (CrypticLib.minecraftVersion() >= 11800) && (platform.equals(IPlatform.Platform.PAPER) || platform.equals(IPlatform.Platform.FOLIA));
     }
 
     private static void saveDefConfigFile(List<File> allFiles) {
@@ -382,14 +420,14 @@ public class RecipeManager {
         Craftorithm.getInstance().saveResource("recipes/example_smithing.yml", false);
         Craftorithm.getInstance().saveResource("recipes/example_stone_cutting.yml", false);
         Craftorithm.getInstance().saveResource("recipes/example_random_cooking.yml", false);
-        Craftorithm.getInstance().saveResource("recipes/example_anvil.yml", false);
+        Craftorithm.getInstance().saveResource("recipes/example_potion.yml", false);
         allFiles.add(new File(recipeFileFolder, "example_shaped.yml"));
         allFiles.add(new File(recipeFileFolder, "example_shapeless.yml"));
         allFiles.add(new File(recipeFileFolder, "example_cooking.yml"));
         allFiles.add(new File(recipeFileFolder, "example_smithing.yml"));
         allFiles.add(new File(recipeFileFolder, "example_stone_cutting.yml"));
         allFiles.add(new File(recipeFileFolder, "example_random_cooking.yml"));
-        allFiles.add(new File(recipeFileFolder, "example_anvil.yml"));
+        allFiles.add(new File(recipeFileFolder, "example_potion.yml"));
     }
 
 }
